- recreate liquibase-phase1.xml
- lazyDatamodels modifications

- Edit and Remove Tasks

+++ src/main/java/com/research/service/impl/DocsServiceImpl.java
@@ -76,9 +76,9 @@ public class DocsServiceImpl extends BaseServiceImpl<Docs> implements DocsServic
 	}
 
 	@Override
-	public int getCountByProject(Long id) {
+	public Long getCountByProject(Long id) {
 		Long count = docsRepository.countByProject(id);
-		return count.intValue();
+		return count;
 	}
 
 }
